Add an option to specify components to DependencyUpdateCheck
For now, the full functionality is not implemented, the controller only
displays a message and exits when encountering DependencyUpdateCheck
with specified components. The controller logic will be implemented as a
part of the next Jira.

Components are specified in a nested structure of workspace name -
application name - component name. It is possible to omit application
name or component name, in which case all applications/components in
a given workspace/application will be processed.

Please note that some changes are autogenerated by kubebuilder.