Run MintMaker only on specified components
Implement an option to only run MintMaker on the components specified in
the dependencyupdatecheck CR. Components are specified in a nested
structure of workspace->application->component. MintMaker will process
all components of a given workspace/application if no lower level
entries are specified. The functionality to filter processed components
is enabled only if the CR contains any components filters, so the
default behavior of MintMaker remains unchanged.