Implement new PipelineRuns controller
This commit implements a new controller/reconciler for PipelineRuns.
Its main responsibility is to launch new pipelineruns when another one
finishes, and try to keep up to a maximum number of pipelineruns active
at any given time.

Besides that, it also launches github pipelineruns first, as a temporary
means of decreasing the chance that their temporary token will expire.