Update Renovate command and set ownership
Changing the placeholder command to the actual renovate command.
Additionally with this commit we are setting all the necessary
ownership to all the resources (ConfigMaps and secrets). So that we
can be sure that one a PipelineRun is removed from the cluster all
its corresponding resources are removed with it.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>