Introduce OSV database to the PipelineRuns
PipelineRuns will now have two tasks and a shared persistent volume. The
first task copies the OSV DB from the periodically built DB image to the
shared volume. The second task runs renovate and can utilize the OSV
database in its run. New env variables were added to specify the DB
location and disable the download of the upstream DB. The created
persistent volume is deleted along with the PipelineRun.

I have increased the memory requirements to 2Gi because I ran into OOM issues
when testing with 1Gi of memory. The reason is most likely that the DB
is loaded into memory during execution, which increases the memory
foorprint by up to ~500MB. It's possible that 2Gi is an overkill, but I
haven't tested the safe threshold more closely.