Support RPMs that require a subscription
RPMs requiring a subscription can be accessed by specifying a cert+key
generated by the subscription manager. There is an existing Konflux
functionality that allows users to build images with RPMs that need a
subscription[1]. Users need to add a secret to Konflux with the
activation key that's used to generate the cert+key. MintMaker can use
the same secret for its own purposes.

The general workflow is:
1. MintMaker controller searches for the activation key secret in user
   namespace
2. If found, it generates the same secret for the PipelineRun
3. PipelineRun uses the secret to generate cert+key
4. Cert+key are placed in a predictable location set to ENV variables
5. ENV variables propagate to the repo file, allowing
   rpm-lockfile-prototype to utilize the cert+key

The generation of cert+key requires root user, so this step is run as
root. Other steps are still run as regular users and even the root
container is not elevated, so there shouldn't be any major security
concerns. Anyuid SCC was added tp tje service account cluster role, sp
that it would have the permission to run containers as root. Specifying
SeccompProfile in the task had to be removed, because it's seemingly
disallowed with the anyuid SCC and the PipelineRun would fail to run. It
shouldn't make a difference since the default profile is used anyway.

RPM repo file set sslclientkey and sslclientcert parameters to locations
of the aforementioned cert+key[2]. Since the processing happens in a
MintMaker pod, users cannot know what paths to set these parameters to.
Thus can be overcome by setting them to ENV variables $SSL_CLIENT_KEY
and $SSL_CLIENT_CERT. MintMaker itself sets those variables to valid
paths, allowing the authentication to happen. THe prefix DNF_VAR_ is
stripped by dnf. To take advantage of this functionality, users will
need to be told to change the paths in their repo files to these
variables.

There is a limitation to this solution: For now, MintMaker can only work
with the default activation key secret. Custom, component-scoped secrets
can be used[3], but it's difficult for mintMaker to determine which
secret belongs to which repo.

[1] https://konflux.pages.redhat.com/docs/users/how-tos/configuring/activation-keys-subscription.html
[2] https://dnf.readthedocs.io/en/latest/conf_ref.html#repo-variables
[3] https://konflux.pages.redhat.com/docs/users/how-tos/configuring/activation-keys-subscription.html#_adding_subscription_entitlement_for_a_specific_component_build