Increase cache times of repo cache and introduce config.js
In order to improve performance and reduce the rate limits, we should
increase the package cache expiration. The duration is set to 4 hours,
which represents one run of MintMaker. I think that having a slightly
stale data is acceptable, since MintMaker runs only every 4 hours and
only updates dependencies once a week for most users.

Self hosted config options can only be written in the config.js
file and will be ignored in renovate.json. To resolve this, transform
the mounted global config file to config.js. A new file, self_hosted.json,
which holds only the self-hosted configuration options was created.
Options that are already in use were transferred from env variables to
this file as well. The file is combined with renovate.json and
mounted to renovate as config.js.

During testing, some cache sets didn't respect the new values. This
is due to how renovate was written, where the cache scaffolding is
sometimes not used and so the configuration is ignored. Renovate's code
needs to be modified to be able to tweak these values.